requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# Computer Vision and ML
ultralytics>=8.0.0  # YOLOv8
albumentations>=1.0.0
scikit-learn>=1.0.0

# Data handling
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Configuration and utilities
PyYAML>=5.4.0
tqdm>=4.60.0
requests>=2.25.0

# Jetson Nano specific (optional, install manually if needed)
# jetson-stats  # For monitoring Jetson performance
# jtop          # Alternative Jetson monitoring

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# Deployment
setuptools>=50.0.0
wheel>=0.36.0
